backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pdfplumber
google-generativeai